google_cloud_challenge/google_cloud_challenge/requirements.txt
streamlit==1.38.0
numpy==2.1.2
pdfplumber==0.11.4
Pillow==10.4.0
python-dotenv==1.0.1
plotly==5.24.1 